Fix the issue with handling decision/fork nodes
The converter worked only if there were multiple fork/decision/case
nodes. The issue was that the return type used to be list if there are
more than 1 element but it was dictionary if there was a single element.

To fix this added a method, which ensures to make the return type always
a list.